+++ jest.config.cjs
@@ -10,4 +10,12 @@ module.exports = {
     ],
   },
   collectCoverageFrom: ["src/**/*.ts"],
+  coverageThreshold: {
+    global: {
+      statements: 100,
+      branches: 100,
+      functions: 100,
+      lines: 100,
+    },
+  },
 };